trading-engine/requirements.txt
ccxt==4.2.0
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
loguru==0.7.2
pydantic==2.5.0
# Backtesting
backtrader==1.9.77.123
vectorbt==0.25.5
# Technical Analysis
ta-lib==0.4.28
# Data visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0
# Database (optional)
sqlalchemy==2.0.23
# Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
